+++ src/app/cotizar/caucion/page.tsx
@@ -1,0 +1,5 @@
+import FormCaucion from '@/components/cotizar/seguros/form-caucion'
+
+export default function Page() {
+  return <FormCaucion />
+}